Darwin fix libstdc++ build warnings.
GCC does not export construction vtable symbols from shared libraries.

The symbols are marked hidden in the objects; for Darwin that makes
them also external (“private_extern” is Darwin’s hidden) which means
that they show up in the list of possible symbols for export from
libstdc++, and there are sufficiently relaxed match conditions that
they reach the exports list.   When Darwin’s static linker encounters
them it generates a warning that they cannot be exported.  This patch
prunes them from the list of symbols to be considered, thus
eliminating the warnings. No functional Change inended to the library
exports.

2018-12-06  Iain Sandoe  <iain@sandoe.co.uk>

	* scripts/make_exports.pl (check names): Don’t try to export
	construction vtable symbols.




git-svn-id: svn://gcc.gnu.org/svn/gcc/trunk@266864 138bc75d-0d04-0410-961f-82ee72b054a4