require profiling support for gcc.dg/lto/20100430-1_0.c test
This test currently fails unexpectedly if GCC is configured with
--disable-gcov, because it requires -fprofile-arcs. This patch
fixes the issue by requiring profiling support in order to run
this test.

Tested with two compilers, one built with --disable-gcov, resulting
in the test reporting an UNSUPPORTED result; and one built with gcov
support, resulting in 2 PASS tests.

gcc/testsuite/ChangeLog:

        * gcc.dg/lto/20100430-1_0.c: Add dg-require-profiling requirement.



git-svn-id: svn://gcc.gnu.org/svn/gcc/trunk@267055 138bc75d-0d04-0410-961f-82ee72b054a4